Move match_policy_minimum_confidence to helpers/compare
The builtin provider was importing a pure matching-confidence helper
from the playlists controller, reversing the provider/controller
dependency direction. PlaylistController itself never called it -
move it next to TrackMatchConfidence in helpers/compare.py where it
belongs.
